refactor(api): разнос singbox_router.go, tunnels.go, subscription.go — чистые перемещения
Этап C1 реструктуризации backend, дословный AST-перенос:

- singbox_router.go (1610 строк) → _dto.go (44 типа запросов/ответов),
  _rules.go (CRUD правил и rule-set'ов + dat-URL), _outbounds.go
  (композиты, пресеты, списки интерфейсов), _policies.go,
  _staging.go (staging + ANSI-очистка ошибок валидации);
  ядро (313 строк) — handler, Enable/Disable/SwitchMode, settings
- tunnels.go (1477 строк) → _dto.go, _view.go (маппинг статусов и
  сборка DTO/снапшотов), _crud.go (List/Create/Update/.../Traffic);
  ядро (211 строк) — интерфейс TunnelService, handler, сеттеры
- subscription.go (1298 строк) → _dto.go (типы + конверторы),
  _members.go (member-операции + стриминг); ядро (394 строки) —
  handler, CRUD, Refresh

Эквивалентность верифицирована мультимножествами строк кода и
комментариев — расхождений нет. swagger.yaml и schemas.gen.ts
перегенерированы (go generate + npm run gen:api): косметический дрейф
порядка полей + унаследованная несвежесть описаний server-listen.
